Fixed screen artifact issue
Fixed the problem with the white rectangles lingering in different menus.
Also changed how the screen loops pass control to other screens. Before, the mainMenu() function was being called multiple times so each time a different screen was navigated to, another version of mainMenu() was being left on the stack.